rename example to TestSafeAreaViewIOS
