+++ apps/src/tests/TestSafeAreaViewIOS/BottomTabs/BottomTabsComponent.tsx
