+++ apps/src/tests/TestSafeAreaViewIOS/BottomTabs/index.tsx
