+++ apps/src/tests/TestSafeAreaViewIOS/shared/ScrollViewAutomatic.tsx
